BlogNews/requirements.txt
Django==4.0.4
django-allauth==0.51.0
django-axes==5.35.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-summernote==0.8.20.0
mysqlclient==2.1.1
Pillow==9.4.0
requests==2.28.1
Faker==15.3.4
Unidecode~=1.3.6
python-dotenv>=1.0.0
